Massive refactor into one recipe.
The `default` recipe is now the only recipe that does
an install.  It installs from the attributes specified
by the `install_type` top level attribute.  Many defaults
are guessed from this name, but can be overridden in the
attributes hash with the same name.  `jvm_opts` and
`config_data` are required right now...

Not sure how this pattern will play out.  We will see.