Fix not being able to install both agent and idnexer.
The way I had planned to support installing both types
does not work as I expected.  This is a huge mess to compensate
for that.  There is now an `install_types` array that keeps
track of the "types" of installs (e.g. agent, indexer, etc..).
We loop through this array and install the configs and services
for each.  I don't know how to feel about this...  Chef...

fixes #5